refactor: move Spring data related code into its own package
