+++ src/main/java/com/example/demo/infrastructure/spring/data/UserMapper.java
@@ -1,4 +1,4 @@
-package com.example.demo.infrastructure.repository;
+package com.example.demo.infrastructure.spring.data;
 
 import com.example.demo.domain.model.User;
 import com.example.demo.infrastructure.model.UserEntity;
